Refoctor : displayTimeAgo 함수 분리
