+++ src/components/Box/Comment.tsx
@@ -6,6 +6,7 @@ import { Reply } from '@emotion-icons/boxicons-regular/Reply';
 import EditCommentForm from './EditCommentForm';
 import { useQuery } from '@tanstack/react-query';
 import { getProfile } from '../../services/profile';
+import { displayTimeAgo } from '../../utils';
 
 const boxItemCss = {
   wrapper: (reply: boolean) => css`
@@ -79,31 +80,6 @@ type Profile = {
   uid: string;
 };
 
-const displayTimeAgo = (postTimestamp: number): string => {
-  const currentTime = new Date();
-  const postTime = new Date(postTimestamp);
-  const timeDifferenceInSeconds = Math.floor((currentTime.getTime() - postTime.getTime()) / 1000);
-
-  const SECONDS_IN_MINUTE = 60;
-  const SECONDS_IN_HOUR = 3600;
-  const SECONDS_IN_DAY = 86400;
-  const SECONDS_IN_WEEK = 604800;
-  const SECONDS_IN_YEAR = 31536000;
-
-  if (timeDifferenceInSeconds < SECONDS_IN_MINUTE) {
-    return `${timeDifferenceInSeconds}s`;
-  } else if (timeDifferenceInSeconds < SECONDS_IN_HOUR) {
-    return `${Math.floor(timeDifferenceInSeconds / SECONDS_IN_MINUTE)}m`;
-  } else if (timeDifferenceInSeconds < SECONDS_IN_DAY) {
-    return `${Math.floor(timeDifferenceInSeconds / SECONDS_IN_HOUR)}h`;
-  } else if (timeDifferenceInSeconds < SECONDS_IN_WEEK) {
-    return `${Math.floor(timeDifferenceInSeconds / SECONDS_IN_DAY)}d`;
-  } else if (timeDifferenceInSeconds < SECONDS_IN_YEAR) {
-    return `${Math.floor(timeDifferenceInSeconds / SECONDS_IN_WEEK)}w`;
-  } else {
-    return `${Math.floor(timeDifferenceInSeconds / SECONDS_IN_YEAR)}y`;
-  }
-};
 const Comment = ({
   setReplyComment,
   setReplyUser,
