+++ src/utils/displayTimeAgo.ts
@@ -1,0 +1,27 @@
+const displayTimeAgo = (postTimestamp: number): string => {
+  const currentTime = new Date();
+  const postTime = new Date(postTimestamp);
+  const timeDifferenceInSeconds = Math.floor((currentTime.getTime() - postTime.getTime()) / 1000);
+
+  const SECONDS_IN_MINUTE = 60;
+  const SECONDS_IN_HOUR = 3600;
+  const SECONDS_IN_DAY = 86400;
+  const SECONDS_IN_WEEK = 604800;
+  const SECONDS_IN_YEAR = 31536000;
+
+  if (timeDifferenceInSeconds < SECONDS_IN_MINUTE) {
+    return `${timeDifferenceInSeconds}s`;
+  } else if (timeDifferenceInSeconds < SECONDS_IN_HOUR) {
+    return `${Math.floor(timeDifferenceInSeconds / SECONDS_IN_MINUTE)}m`;
+  } else if (timeDifferenceInSeconds < SECONDS_IN_DAY) {
+    return `${Math.floor(timeDifferenceInSeconds / SECONDS_IN_HOUR)}h`;
+  } else if (timeDifferenceInSeconds < SECONDS_IN_WEEK) {
+    return `${Math.floor(timeDifferenceInSeconds / SECONDS_IN_DAY)}d`;
+  } else if (timeDifferenceInSeconds < SECONDS_IN_YEAR) {
+    return `${Math.floor(timeDifferenceInSeconds / SECONDS_IN_WEEK)}w`;
+  } else {
+    return `${Math.floor(timeDifferenceInSeconds / SECONDS_IN_YEAR)}y`;
+  }
+};
+
+export default displayTimeAgo;
