+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export { default as errorObjToString } from './errorObjToString';
 export { default as extractUsernameFromEmail } from './extractUsernameFromEmail';
 export { default as requiredFormValue } from './requiredFormValue';
+export { default as displayTimeAgo } from './displayTimeAgo';
